Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Please test the outputted image by jenkins and let me know when to run a prod release.

I've tested the image outputted by jenkins, It boots but it has the same issues as the old device tree (The one made by shakalaca). It's like it's not using the latest dt. But I might have got the wrong image download link since I'm not familiarized with navigating on the jenkins website. If you can, please send me the correct link (if I got the wrong one) to download the test img.

Patch-set: 1